requirements.txt
ddt==1.2.1
gitdb==0.6.4
gitdb2==2.0.5
GitPython==2.1.12
PyQt5==5.13.0
PyQt5-sip==4.19.18
PyYAML==5.1.1
smmap==0.9.0
smmap2==2.0.5